+++ src/Raven.CodeAnalysis/SynthesizedUnionMethodNames.cs
@@ -4,4 +4,5 @@ internal static class SynthesizedUnionMethodNames
 {
     public const string DisplayNameHelper = "<RavenUnionDisplayName>";
     public const string FriendlyTypeNameHelper = "<RavenFriendlyTypeName>";
+    public const string FormatValueHelper = "<RavenFormatUnionValue>";
 }